+++ comb_spec_searcher/strategies/constructor.py
@@ -54,7 +54,7 @@ class Constructor(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectTy
     @abc.abstractmethod
     def get_sub_objects(
         self, subgens: SubGens, n: int, **parameters: int
-    ) -> Iterator[Tuple[CombinatorialObjectType, ...]]:
+    ) -> Iterator[Tuple[Optional[CombinatorialObjectType], ...]]:
         """Return the subobjs/image of the bijection implied by the constructor."""
 
     @abc.abstractmethod
@@ -262,12 +262,17 @@ class CartesianProduct(Constructor[CombinatorialClassType, CombinatorialObjectTy
     def get_sub_objects(
         self, subgens: SubGens, n: int, **parameters: int
     ) -> Iterator[Tuple[CombinatorialObjectType, ...]]:
-        assert len(parameters) == 0, "only implemented in one variable, namely 'n'"
-        for comp in self._valid_compositions(n):
-            for sub_objs in product(
-                *tuple(subgen(n=d["n"]) for d, subgen in zip(comp, subgens))
+        for child_parameters in self._valid_compositions(n, **parameters):
+            extra_parameters = self.get_extra_parameters(child_parameters)
+            if extra_parameters is None:
+                continue
+            for objs in product(
+                *[
+                    gen(n=extra_params.pop("n"), **extra_params)
+                    for gen, extra_params in zip(subgens, extra_parameters)
+                ]
             ):
-                yield tuple(sub_objs)
+                yield tuple(objs)
 
     def random_sample_sub_objects(
         self,
@@ -277,19 +282,24 @@ class CartesianProduct(Constructor[CombinatorialClassType, CombinatorialObjectTy
         n: int,
         **parameters: int
     ):
-        assert not parameters, "only implemented in one variable"
         random_choice = randint(1, parent_count)
         total = 0
-        for comp in self._valid_compositions(n):
+        for child_parameters in self._valid_compositions(n, **parameters):
             tmp = 1
-            for i, rec in enumerate(subrecs):
-                tmp *= rec(n=comp[i]["n"])
+            extra_parameters = self.get_extra_parameters(child_parameters)
+            if extra_parameters is None:
+                continue
+            for rec, extra_params in zip(subrecs, extra_parameters):
+                tmp *= rec(n=extra_params.pop("n"), **extra_params)
                 if tmp == 0:
                     break
             total += tmp
             if random_choice <= total:
+                extra_parameters = self.get_extra_parameters(child_parameters)
+                assert extra_parameters is not None
                 return tuple(
-                    subsampler(d["n"]) for d, subsampler in zip(comp, subsamplers)
+                    subsampler(n=extra_params.pop("n"), **extra_params)
+                    for subsampler, extra_params in zip(subsamplers, extra_parameters)
                 )
 
     @staticmethod
@@ -399,19 +409,23 @@ class DisjointUnion(Constructor[CombinatorialClassType, CombinatorialObjectType]
             res += rec(n=n, **extra_params)
         return res
 
-    @staticmethod
     def get_sub_objects(
-        subgens: SubGens, n: int, **parameters: int
-    ) -> Iterator[Tuple[CombinatorialObjectType, ...]]:
-        assert len(parameters) == 0, "only implemented in one variable, namely 'n'"
-        for i, subgen in enumerate(subgens):
-            for gp in subgen(n, **parameters):
-                yield tuple(None for _ in range(i)) + (gp,) + tuple(
-                    None for _ in range(len(subgens) - i - 1)
+        self, subgens: SubGens, n: int, **parameters: int
+    ) -> Iterator[Tuple[Optional[CombinatorialObjectType], ...]]:
+        for (idx, subgen), extra_params in zip(
+            enumerate(subgens), self.get_extra_parameters(n, **parameters)
+        ):
+            if extra_params is None or any(
+                val != 0 and k in self.zeroes[idx] for k, val in parameters.items()
+            ):
+                continue
+            for gp in subgen(n, **extra_params):
+                yield tuple(None for _ in range(idx)) + (gp,) + tuple(
+                    None for _ in range(len(subgens) - idx - 1)
                 )
 
-    @staticmethod
     def random_sample_sub_objects(
+        self,
         parent_count: int,
         subsamplers: SubSamplers,
         subrecs: SubRecs,
@@ -420,10 +434,18 @@ class DisjointUnion(Constructor[CombinatorialClassType, CombinatorialObjectType]
     ):
         random_choice = randint(1, parent_count)
         total = 0
-        for idx, (subrec, subsampler) in enumerate(zip(subrecs, subsamplers)):
-            total += subrec(n=n, **parameters)
+        for (idx, rec), subsampler, extra_params in zip(
+            enumerate(subrecs), subsamplers, self.get_extra_parameters(n, **parameters)
+        ):
+            # if a parent parameter is not mapped to by some child parameter
+            # then it is assumed that the value of the parent parameter must be 0
+            if extra_params is None or any(
+                val != 0 and k in self.zeroes[idx] for k, val in parameters.items()
+            ):
+                continue
+            total += rec(n=n, **extra_params)
             if random_choice <= total:
-                obj = subsampler(n=n, **parameters)
+                obj = subsampler(n=n, **extra_params)
                 return (
                     tuple(None for _ in range(idx))
                     + (obj,)